+++ src/static/dolphin/exi/EXIUart.c
@@ -111,7 +111,7 @@ int ReadUARTN() {
 }
 
 static int QueueLength(void) {
-    unsigned long cmd;
+    u32 cmd;
 
     if (EXISelect(Chan, Dev, 3) == 0) {
         return -1;
@@ -126,8 +126,8 @@ static int QueueLength(void) {
 }
 
 int WriteUARTN(void *buf, u32 len) {
-    unsigned long cmd;
-    long xLen;
+    u32 cmd;
+    s32 xLen;
     int qLen;
     char* ptr;
     int locked;
@@ -175,7 +175,7 @@ int WriteUARTN(void *buf, u32 len) {
                     break;
                 }
 
-                xLen = len < 4 ? (long)len : 4;
+                xLen = len < 4 ? (s32)len : 4;
                 
                 EXIImm(Chan, buf, xLen, 1, 0);
                 (char*)buf += xLen;
